+++ packages/canonical-bridge-widget/src/core/utils/string.ts
@@ -1,5 +1,6 @@
 import { MIN_FEE } from '@/core/constants';
 import { formatNumber } from '@/core/utils/number';
+import { IFeeDetails } from '@/modules/aggregator';
 
 export function truncateStr(str: string, headLen = 6, tailLen = 6) {
   if (!str) {
@@ -44,6 +45,22 @@ export function formatFeeAmount(amount: string | number) {
   }
 }
 
+export function formatRouteFees(feeList: IFeeDetails[]) {
+  const result = feeList.reduce((acc: { [key: string]: number }, item) => {
+    const symbol = item.symbol;
+    const value = Number(item.value);
+    if (symbol && !acc[symbol]) acc[symbol] = 0;
+    acc[symbol] += value;
+    return acc;
+  }, {} as { [key: string]: number });
+  const resultString = Object.keys(result)
+    .map((symbol) => {
+      return `${formatFeeAmount(result[symbol])} ${symbol}`;
+    })
+    .join(' + ');
+  return resultString ? resultString : '';
+}
+
 export function utf8ToHex(utf8Str: any) {
   return Array.from(utf8Str)
     .map((char: any) => char.charCodeAt(0).toString(16).padStart(2, '0'))
@@ -52,4 +69,9 @@ export function utf8ToHex(utf8Str: any) {
 
 export function includesIgnoreCase(strArr: string[], target: string) {
   return !!strArr?.find((e) => e.toUpperCase() === target?.toUpperCase());
+}
+
+export function capitalizeFirst(text: string) {
+  if (typeof text !== 'string') return '';
+  return text.charAt(0).toUpperCase() + text.slice(1).toLowerCase();
 }