+++ packages/canonical-bridge-widget/src/modules/transfer/hooks/useToTokenInfo.ts
@@ -10,7 +10,7 @@ export const useToTokenInfo = () => {
     return {
       cBridge: toTokenInfo?.cBridge?.raw?.token.decimal || 0,
       deBridge: toTokenInfo?.deBridge?.raw?.decimals || 0,
-      stargate: selectedToken?.stargate?.raw?.decimals || 0,
+      stargate: selectedToken?.stargate?.raw?.token?.decimals || 0,
       layerZero: selectedToken?.layerZero?.raw?.decimals || 0,
       meson: selectedToken?.meson?.raw?.decimals || 0,
     };
@@ -19,7 +19,7 @@ export const useToTokenInfo = () => {
     return {
       cBridge: toTokenInfo?.cBridge?.raw?.token.address || '',
       deBridge: toTokenInfo?.deBridge?.raw?.address || '',
-      stargate: toTokenInfo?.stargate?.raw?.address || '',
+      stargate: toTokenInfo?.stargate?.raw?.token?.address || '',
       layerZero: toTokenInfo?.layerZero?.raw?.address || '',
       meson: toTokenInfo?.meson?.raw?.addr || 0,
     };
@@ -29,14 +29,14 @@ export const useToTokenInfo = () => {
     return {
       cBridge: toTokenInfo?.cBridge?.raw?.token.symbol || '',
       deBridge: toTokenInfo?.deBridge?.raw?.symbol || '',
-      stargate: toTokenInfo?.stargate?.raw?.symbol || '',
+      stargate: toTokenInfo?.stargate?.raw?.token?.symbol || '',
       layerZero: toTokenInfo?.layerZero?.raw?.symbol || '',
       meson: toTokenInfo?.meson?.raw?.id?.toUpperCase() || '',
     };
   }, [
     toTokenInfo?.cBridge?.raw?.token.symbol,
     toTokenInfo?.deBridge?.raw?.symbol,
-    toTokenInfo?.stargate?.raw?.symbol,
+    toTokenInfo?.stargate?.raw?.token?.symbol,
     toTokenInfo?.layerZero?.raw?.symbol,
     toTokenInfo?.meson?.raw?.id,
   ]);